Feature/wa 271 todo fixes

+++ IdentityAccess/IdentityAccessImplementations/persistence/InMemoryUserRepositoryTests.swift
@@ -8,23 +8,31 @@ import IdentityAccessDomainModel
 
 class InMemoryUserRepositoryTests: XCTestCase {
 
-    let repository: UserRepository = InMemoryUserRepository()
+    let repository: SingleUserRepository = InMemoryUserRepository()
     var user: User!
     var other: User!
 
     override func setUp() {
         super.setUp()
         DomainRegistry.put(service: MockEncryptionService(), for: EncryptionServiceProtocol.self)
-        DomainRegistry.put(service: repository, for: UserRepository.self)
+        DomainRegistry.put(service: repository, for: SingleUserRepository.self)
         DomainRegistry.put(service: MockBiometricService(), for: BiometricAuthenticationService.self)
         DomainRegistry.put(service: IdentityService(), for: IdentityService.self)
         do {
-            user = try DomainRegistry.identityService.registerUser(password: "Mypass123")
-            try repository.remove(user)
-            other = try DomainRegistry.identityService.registerUser(password: "Otherpass123")
-            try repository.remove(other)
+            _ = try DomainRegistry.identityService.registerUser(password: "Mypass123")
+            user = repository.primaryUser()!
+            try removePrimaryUser()
+            _ = try DomainRegistry.identityService.registerUser(password: "Otherpass123")
+            other = repository.primaryUser()
+            try removePrimaryUser()
         } catch {
             XCTFail("Failed to setUp")
+        }
+    }
+
+    private func removePrimaryUser() throws {
+        if let user = repository.primaryUser() {
+            try repository.remove(user)
         }
     }
